Fix TestNativeWindowConstructor to use proper Mocks.
The SimpleNamespaces it was using before were
not actually working. The mistake was hidden
on Windows due to NativeWindow falling back
to using _NativeWindowWin32.

To avoid this kind of issue in the future,
NativeWindow will now fall back to the platform
default only if no supported _osNativeWindow
argument is actually provided. It will not
swallow exceptions any more.