Move various classes from package 'view' to 'model'.
The classes for retrieving and modifying native
windows in the operating system were initially
miscategorized as being part of the 'view' in
the model-view-controller architecture. Since the
whole point of the program is manipulating
programs and windows, and no UI code is directly
using those classes, they belong in the model
package.