Re-implement TestWindowCreator using Tkinter instead of wxPython.
Using wxPython was causing two main issues.

First, windows were not actually being destroyed
correctly when the destroy_all() method was
called. The needed method to correct that bug
was not yet implemented in the Phoenix version
of wxPython.

Second, wxPython's AssociateHandle() method could
not be used on windows created by
TestWindowCreator when wxPython was also the one
creating the windows.